modifying

+++ transaction-propogation/transaction-propogation/src/main/java/sdj/repository/LogRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import sdj.entity.Log;
 
-public interface LogRepository extends JpaRepository<Log, Long> {
+public interface LogRepository extends JpaRepository<Log, Integer> {
 	
 }
